fix: minor folder refactor
