+++ src/http/routes/admin/jwks.ts
@@ -2,7 +2,7 @@ import { FastifyInstance, RequestGenericInterface } from 'fastify'
 import apiKey from '../../plugins/apikey'
 import { jwksManager } from '@internal/database'
 import { FromSchema } from 'json-schema-to-ts'
-import { UrlSigningJwkGenerator } from '@internal/auth/generators/jwk-generator'
+import { UrlSigningJwkGenerator } from '@internal/auth/jwks/generator'
 import { logSchema } from '@internal/monitoring'
 
 const addSchema = {
