--- src/internal/database/jwks-manager/index.ts
@@ -1,1 +1,0 @@
-export * from './manager'
